fix parsing of forms
