+++ web/marketing/pages/free-offers-modal.html
@@ -29,8 +29,8 @@ See LICENSE for copying information. -->
                     </div>
                     <div class="form-row">
                         <div class="form-group col-md-4">
-                            <label for="AwardCreditInCents">Award Credit</label>
-                            <input type="number" class="form-control" name="AwardCreditInCents" id="AwardCreditInCents" placeholder="$50" min="1" required>                        
+                            <label for="AwardCredit">Award Credit</label>
+                            <input type="number" class="form-control" name="AwardCredit" id="AwardCredit" placeholder="$50" min="1" required>                        
                         </div>
                         <div class="form-group col-md-4">
                             <label for="AwardCreditDurationDays">Award Credit Exp</label>
