+++ web/marketing/pages/referral-offers-modal.html
@@ -29,16 +29,16 @@ See LICENSE for copying information. -->
                     </div>
                     <div class="form-row">
                         <div class="form-group col-md-2">
-                            <label for="InviteeCreditInCents">Give Credit</label>
-                            <input type="number" class="form-control" name="InviteeCreditInCents" id="InviteeCreditInCents" placeholder="$50" required>                        
+                            <label for="InviteeCredit">Give Credit</label>
+                            <input type="number" class="form-control" name="InviteeCredit" id="InviteeCredit" placeholder="$50" required>                        
                         </div>
                         <div class="form-group col-md-2">
                             <label for="InviteeCreditDurationDays">Give Credit Exp.</label>
                             <input type="number" class="form-control" name="InviteeCreditDurationDays" id="InviteeCreditDurationDays" min="1" placeholder="14 days" required>
                         </div>
                         <div class="form-group col-md-2">
-                            <label for="AwardCreditInCents">Award Credit</label>
-                            <input type="number" class="form-control" name="AwardCreditInCents" id="AwardCreditInCents" placeholder="$50" min="1" required>                        
+                            <label for="AwardCredit">Award Credit</label>
+                            <input type="number" class="form-control" name="AwardCredit" id="AwardCredit" placeholder="$50" min="1" required>                        
                         </div>
                         <div class="form-group col-md-3">
                             <label for="AwardCreditDurationDays">Award Credit Exp.</label>
